move 'cancel' directory to 'pages' directory
