+++ src/applications/vaos/appointment-list/pages/RequestedAppointmentDetailsPage/RequestedAppointmentDetailsPage.jsx
@@ -13,14 +13,14 @@ import VAFacilityLocation from '../../../components/VAFacilityLocation';
 import { selectFeatureBreadcrumbUrlUpdate } from '../../../redux/selectors';
 import { FETCH_STATUS } from '../../../utils/constants';
 import { scrollAndFocus } from '../../../utils/scrollAndFocus';
-import CancelConfirmationPage from '../../components/cancel/CancelConfirmationPage';
-import CancelWarningPage from '../../components/cancel/CancelWarningPage';
 import PageLayout from '../../components/PageLayout';
 import {
   closeCancelAppointment,
   fetchRequestDetails,
 } from '../../redux/actions';
 import { selectRequestedAppointmentDetails } from '../../redux/selectors';
+import CancelConfirmationPage from '../CancelAppointmentPage/CancelConfirmationPage';
+import CancelWarningPage from '../CancelAppointmentPage/CancelWarningPage';
 
 export default function RequestedAppointmentDetailsPage() {
   const { id } = useParams();
